Rename assessment files
